yuk/requirements.txt
Django==1.3.1
South==0.7.3
amqplib==1.0.2
anyjson==0.3.1
celery==2.4.0
django-extensions==0.7.1
django-haystack==1.2.5
django-registration
django-taggit==0.9.3
django-taggit-templatetags==0.4.6dev
django-templatetag-sugar==0.1
git+git://github.com/ipython/ipython.git
kombu==1.4.3
python-dateutil==1.5
wsgiref==0.1.2
